requirements.txt
APScheduler==3.6.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.1
h11==0.9.0
httptools==0.1.1
idna==2.10
picamera==1.13
pydantic==1.6.1
PyJWT==1.7.1
pytz==2020.1
qiniu==7.2.10
requests==2.24.0
six==1.15.0
starlette==0.13.6
tzlocal==2.1
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
